Cambio de Data Type de id a Long, endpoint de listar Usuarios

+++ src/main/java/com/alkemy/cysjava/virtualwallet/DTOs/UserDTO.java
@@ -1,5 +1,6 @@
 package com.alkemy.cysjava.virtualwallet.DTOs;
 
+import jakarta.validation.constraints.Null;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserDTO {
-    private int id;
+    private Long id;
     private String firstname;
     private String lastname;
     private String email;
